+++ accommodationservice/src/main/java/com/megatravel/controllers/AccommodationTypeController.java
@@ -1,5 +1,7 @@
 package com.megatravel.controllers;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.megatravel.decodeJWT.DecodeJwtToken;
 import com.megatravel.dtosoap.hotel.AccomodationTypeDTO;
 import com.megatravel.services.AccommodationTypeService;
 
@@ -23,22 +26,41 @@ public class AccommodationTypeController {
 	private AccommodationTypeService service;
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<AccomodationTypeDTO> getRoomType(@PathVariable("id") Long id, @PathVariable("room-id") Long room) {
+	public ResponseEntity<AccomodationTypeDTO> getRoomType(@PathVariable("id") Long id, @PathVariable("room-id") Long room,
+			HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("getRoomType", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AccomodationTypeDTO>(service.getRoomType(id), HttpStatus.OK);
 	}
 	
 	@RequestMapping(method = RequestMethod.POST)
-	public ResponseEntity<AccomodationTypeDTO> createAccommodationType(@RequestBody AccomodationTypeDTO type, @PathVariable("room-id") Long id) {
+	public ResponseEntity<AccomodationTypeDTO> createAccommodationType(@RequestBody AccomodationTypeDTO type, @PathVariable("room-id") Long id,
+			HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("createAccommodationType", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AccomodationTypeDTO>(service.createAccommodationType(type), HttpStatus.CREATED);
 	}
 	
 	@RequestMapping(method = RequestMethod.PUT)
-	public ResponseEntity<AccomodationTypeDTO> updateAccommodationType(@RequestBody AccomodationTypeDTO type, @PathVariable("room-id") Long id) {
+	public ResponseEntity<AccomodationTypeDTO> updateAccommodationType(@RequestBody AccomodationTypeDTO type, @PathVariable("room-id") Long id,
+			HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("updateAccommodationType", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AccomodationTypeDTO>(service.updateAccommodationType(type), HttpStatus.ACCEPTED);
 	}
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Boolean> removeAccommodationType(@PathVariable("id") Long id) {
+	public ResponseEntity<Boolean> removeAccommodationType(@PathVariable("id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("removeAccommodationType", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<Boolean>(service.removeAccommodationType(id),HttpStatus.ACCEPTED);
 	}
