+++ accommodationservice/src/main/java/com/megatravel/controllers/AddressController.java
@@ -1,5 +1,7 @@
 package com.megatravel.controllers;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.megatravel.decodeJWT.DecodeJwtToken;
 import com.megatravel.dtosoap.global_parameters.AddressDTO;
 import com.megatravel.services.AddressService;
 
@@ -22,22 +25,38 @@ public class AddressController {
 	private AddressService service;
 	
 	@RequestMapping(method = RequestMethod.GET)
-	public ResponseEntity<AddressDTO> getHotelsAddress(@PathVariable("hotel-id") Long id) {
+	public ResponseEntity<AddressDTO> getHotelsAddress(@PathVariable("hotel-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("getHotelsAddress", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AddressDTO>(service.getHotelsAddress(id), HttpStatus.OK);
 	}
 	
 	@RequestMapping(method = RequestMethod.POST)
-	public ResponseEntity<AddressDTO> createAddress(@RequestBody AddressDTO address, @PathVariable("hotel-id") Long id) {
+	public ResponseEntity<AddressDTO> createAddress(@RequestBody AddressDTO address, @PathVariable("hotel-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("createAddress", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AddressDTO>(service.createAddress(address), HttpStatus.CREATED);
 	}
 	
 	@RequestMapping(method = RequestMethod.PUT)
-	public ResponseEntity<AddressDTO> updateAddress(@RequestBody AddressDTO address, @PathVariable("hotel-id") Long id) {
+	public ResponseEntity<AddressDTO> updateAddress(@RequestBody AddressDTO address, @PathVariable("hotel-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("updateAddress", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<AddressDTO>(service.updateAddress(address), HttpStatus.ACCEPTED);
 	}
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Boolean> removeAddress(@PathVariable("id") Long id) {
+	public ResponseEntity<Boolean> removeAddress(@PathVariable("id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("removeAddress", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<Boolean>(service.removeAddress(id),HttpStatus.ACCEPTED);
 	}
