+++ accommodationservice/src/main/java/com/megatravel/controllers/ExtraOptionController.java
@@ -2,6 +2,8 @@ package com.megatravel.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.megatravel.decodeJWT.DecodeJwtToken;
 import com.megatravel.dtosoap.hotel.ExtraOptionDTO;
 import com.megatravel.services.ExtraOptionService;
 
@@ -24,32 +27,58 @@ public class ExtraOptionController {
 	private ExtraOptionService service;
 	
 	@RequestMapping(value="/hotel/{hotel-id}",method = RequestMethod.GET)
-	public ResponseEntity<List<ExtraOptionDTO>> getRoomExtraOptionsWithHotelId(@PathVariable("hotel-id") Long id) {
+	public ResponseEntity<List<ExtraOptionDTO>> getRoomExtraOptionsWithHotelId(@PathVariable("hotel-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("getRoomExtraOptionsWithHotelId", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<List<ExtraOptionDTO>>(service.getHotelExtraOption(id), HttpStatus.OK);
 	}
 	
 	@RequestMapping(value="/room/{room-id}",method = RequestMethod.GET)
-	public ResponseEntity<List<ExtraOptionDTO>> getRoomExtraOptionsWithRoomId(@PathVariable("room-id") Long id) {
+	public ResponseEntity<List<ExtraOptionDTO>> getRoomExtraOptionsWithRoomId(@PathVariable("room-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("getRoomExtraOptionsWithRoomId", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<List<ExtraOptionDTO>>(service.getRoomExtraOptions(id), HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<ExtraOptionDTO> getRoomExtraOption(@PathVariable("id") Long id) {
+	public ResponseEntity<ExtraOptionDTO> getRoomExtraOption(@PathVariable("id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("getRoomExtraOption", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<ExtraOptionDTO>(service.getExtraOption(id), HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/room/{room-id}", method = RequestMethod.POST)
-	public ResponseEntity<ExtraOptionDTO> createExtraOption(@RequestBody ExtraOptionDTO extraOption, @PathVariable("room-id") Long id) {
+	public ResponseEntity<ExtraOptionDTO> createExtraOption(@RequestBody ExtraOptionDTO extraOption, 
+			@PathVariable("room-id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("createExtraOption", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<ExtraOptionDTO>(service.createRoomExtraOption(extraOption), HttpStatus.CREATED);
 	}
 	
 	@RequestMapping(value = "room/{room-id}",method = RequestMethod.PUT)
-	public ResponseEntity<ExtraOptionDTO> updateRoom(@RequestBody ExtraOptionDTO extraOption,@PathVariable("room-id") Long id) {
+	public ResponseEntity<ExtraOptionDTO> updateRoom(@RequestBody ExtraOptionDTO extraOption,@PathVariable("room-id") Long id
+			, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("updateRoom", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<ExtraOptionDTO>(service.updateRoomExtraOption(extraOption), HttpStatus.ACCEPTED);
 	}
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Boolean> removeRoom(@PathVariable("id") Long id) {
+	public ResponseEntity<Boolean> removeRoom(@PathVariable("id") Long id, HttpServletRequest request) {
+		if(!DecodeJwtToken.canAccessMethod("removeRoom", request.getHeader("Authorization"))) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		
 		return new ResponseEntity<Boolean>(service.removeExtraOption(id),HttpStatus.ACCEPTED);
 	}
 }
